+++ packages/qti-convert-core/src/qti-converter/index.ts
@@ -1,2 +1,3 @@
 export * from './converter/converter';
 export * from '../qti-downgrader';
+export * from '../qti-upgrader';