+++ packages/qti-convert-core/src/qti-upgrader/index.ts
@@ -1,0 +1,1 @@
+export { upgradeQti2toQti3 } from './convert-qti2-to-qti3';